Improve manual alignment mode

+++ src/funcs_zarr.py
@@ -237,7 +237,8 @@ def preallocate_zarr(name, group, dimx, dimy, dimz, dtype, overwrite):
         #     arr.zeros(name=group, shape=shape, chunks=chunkshape, dtype=dtype, overwrite=overwrite)
         # else:
         # arr.zeros(name=group, shape=shape, chunks=chunkshape, dtype=dtype, compressor=compressor, overwrite=overwrite, synchronizer=synchronizer)
-        arr.zeros(name=group, shape=shape, chunks=chunkshape, dtype=dtype, compressor=compressor, overwrite=overwrite)
+        # arr.zeros(name=group, shape=shape, chunks=chunkshape, dtype=dtype, compressor=compressor, overwrite=overwrite)
+        arr.zeros(name=group, shape=shape, chunks=chunkshape, dtype='|u1', compressor=compressor, overwrite=overwrite)
         '''dtype definitely sets the dtype, otherwise goes to float64 on Lonestar6, at least for use with tensorstore'''
         # write_metadata_zarr_multiscale() # thon3 al   write single multiscale zarr for all aligned s
     except:
@@ -245,7 +246,6 @@ def preallocate_zarr(name, group, dimx, dimy, dimz, dtype, overwrite):
         cfg.main_window.warn('Zarr Preallocation Encountered A Problem')
     else:
         cfg.main_window.hud.done()
-        # cfg.main_window.hud(output_text)
         logger.info(output_text)
